+++ src/PrincipalComponents.cxx
@@ -6,7 +6,7 @@ bool ReadMatrixFile(vnl_matrix<double> &mat, const char *file)
   FILE *fin = fopen(file,"rt");
   if(fin == NULL) return false;
 
-  size_t r = 0, c = 0;
+  int r = 0, c = 0;
   fscanf(fin, "# MATRIX %d %d\n", &r, &c);
   if(r == 0 || c == 0)
     { fclose(fin); return false; }
